+++ packages/builder/src/pages/builder/portal/users/_components/ActiveDirectoryInfo.svelte
